친구 상태 확인 버그 수정
