+++ app/(main)/@modal/(.)[nickName]/friends/page.tsx
@@ -3,6 +3,7 @@ import Modal from "@/components/Modal";
 import Friends from "@/features/friends/Friends";
 import { Suspense } from "react";
 import FriendsSkeleton from "@/features/friends/FriendsSkeleton";
+import { getFriends } from "@/lib/getFriends";
 
 export default async function Page({
   params,
@@ -10,11 +11,13 @@ export default async function Page({
   params: Promise<{ nickName: string }>;
 }) {
   const { nickName } = await params;
+  const { friends, user, userCache } = await getFriends(nickName);
+
   return (
-    <Modal title={<h1>친구 목록</h1>}>
-      <div className="mt-8 w-lg overflow-y-auto h-130">
+    <Modal title={<h1 className="text-center">친구</h1>}>
+      <div className="mt-8 w-md overflow-y-auto h-130">
         <Suspense fallback={<FriendsSkeleton />}>
-          <Friends nickName={nickName} />
+          <Friends friends={friends} user={user} userCache={userCache} />
         </Suspense>
       </div>
     </Modal>
