+++ features/friends/Friends.tsx
@@ -1,79 +1,17 @@
 import ProfileImg from "@/components/ProfileImg";
 import { notFound } from "next/navigation";
-import { cookies } from "next/headers";
-import type { User } from "@/types/User";
-import type { FriendData } from "@/types/Friends";
 import Link from "next/link";
+import { FriendData } from "@/types/Friends";
+import { User } from "@/types/User";
 
 type Props = {
-  nickName: string;
+  friends: FriendData[];
+  user: User;
+  userCache: Map<number, User | undefined>;
 };
 
-const userCache = new Map<number, User | undefined>();
-
-export default async function Friends({ nickName }: Props) {
+export default async function Friends({ friends, user, userCache }: Props) {
   try {
-    const cookieStore = await cookies();
-    const token = cookieStore.get("token")?.value;
-
-    async function getFriends() {
-      const res = await fetch(
-        `${process.env.NEXT_PUBLIC_SERVER_URL}/api2/friendships/${nickName}`,
-        {
-          method: "GET",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-      return (await res.json()).data as FriendData[]; // 친구신청 목록 배열
-    }
-
-    async function getUser() {
-      const res = await fetch(
-        `${process.env.NEXT_PUBLIC_SERVER_URL}/api2/members/${nickName}`,
-        {
-          method: "GET",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-      return (await res.json()).data as User; // 해당 페이지의 유저
-    }
-
-    // 이게 최선인가.... 나중에 더 좋은 api가 나오면 바꾸자.......
-    const [friends, user] = await Promise.all([getFriends(), getUser()]);
-
-    await Promise.all(
-      friends.map(async (friend) => {
-        const id =
-          user.id === friend.receiverId
-            ? friend.requesterId
-            : friend.receiverId;
-        if (!userCache.has(id)) {
-          userCache.set(id, undefined);
-          const userRes = await fetch(
-            `${process.env.NEXT_PUBLIC_SERVER_URL}/api/members/${id}`,
-            {
-              method: "GET",
-              headers: {
-                "Content-Type": "application/json",
-                Authorization: `Bearer ${token}`,
-              },
-            }
-          );
-          const userJson = await userRes.json();
-          userCache.set(id, userJson.data);
-          return userJson;
-        }
-      })
-    );
-
-    // const list = Array(10).fill(0);
-    // 나중에 id 하나씩 조회하는 로직 추가해야 함...!!
     return (
       <>
         {friends.length === 0 && (
